+++ Documentation/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Changed
+- Renamed the `-DBPath` parameter into `-DatabasePath`.
+
 ## 3.5.1 - 2019-05-23
 
 This is a [Chocolatey](https://chocolatey.org/packages/dsinternals-psmodule)-only release.